Centralize testing utils
