+++ containers/record-linkage/app/linkage/mpi.py
@@ -5,7 +5,7 @@ import uuid
 from functools import cache
 from typing import Dict
 from typing import List
-from typing import Literal
+from typing import Union
 
 from app.linkage.core import BaseMPIConnectorClient
 from app.linkage.dal import DataAccessLayer
@@ -603,7 +603,7 @@ class DIBBsMPIConnectorClient(BaseMPIConnectorClient):
                 )
 
     @cache
-    def _get_external_source_id(self, external_source_name: str) -> Literal[str, None]:
+    def _get_external_source_id(self, external_source_name: str) -> Union[str, None]:
         """
         Gets the external source id for the external source name provided.
         :param external_source_name: The external source name.
